removed console.logs, unneeded imports, and added in edit and delete modals on call list
